[#0] draw ideas for Calculator structure
