+++ Calculator/Calculator/Calculators.swift
@@ -1,97 +1,97 @@
+////
+////  Calculator.swift
+////  Calculator
+////
+////  Created by Seungjin Baek on 2021/03/25.
+////
 //
-//  Calculator.swift
-//  Calculator
+//import Foundation
 //
-//  Created by Seungjin Baek on 2021/03/25.
+////enum Operator {
+////    case add, multiply, subtract
+////
+////    func operation(firstElement: Int, secondElement: Int) -> Int { // 타입이 반드시 입력되어야함
+////        switch self {
+////        case .add : return firstElement + secondElement
+////        case .multiply : return firstElement * secondElement
+////        case .subtract : return firstElement - secondElement
+////    }
+////  }
+////}
 //
-
-import Foundation
-
-//enum Operator {
-//    case add, multiply, subtract
-//
-//    func operation(firstElement: Int, secondElement: Int) -> Int { // 타입이 반드시 입력되어야함
-//        switch self {
-//        case .add : return firstElement + secondElement
-//        case .multiply : return firstElement * secondElement
-//        case .subtract : return firstElement - secondElement
-//    }
-//  }
+//protocol DecimalOperator {
+//    func multiply(_ firstElement: String, by secondElement: String) -> String
+//    func divide()
 //}
-
-protocol DecimalOperator {
-    func multiply(_ firstElement: String, by secondElement: String) -> String
-    func divide()
-}
-
-protocol BinaryOperator {
-    func andOperation()
-    func nandOperation()
-    func orOperation()
-    func norOperation()
-    func xorOperation()
-    func notOperation()
-    func rightShiftOperation()
-    func leftShiftOPeration()
-}
-
-class Calculator {
-    func add(_ firstElement: String, to secondElement: String) {
-        
-    }
-    
-    func subtract(_ secondElement: String, from firstElement: String) {
-        
-    }
-    
-    func clear() {
-        
-    }
-}
-
-class DecimalCalculator: Calculator, DecimalOperator {
-    func multiply(_ firstElement: String, by secondElement: String) -> String {
-        return ""
-    }
-    
-    func divide() {
-        
-    }
-}
-
-class BinaryCalculator: Calculator, BinaryOperator {
-    func andOperation() {
-        
-    }
-    
-    func nandOperation() {
-        
-    }
-    
-    func orOperation() {
-        
-    }
-    
-    func norOperation() {
-        
-    }
-    
-    func xorOperation() {
-        
-    }
-    
-    func notOperation() {
-        
-    }
-    
-    func rightShiftOperation() {
-        
-    }
-    
-    func leftShiftOPeration() {
-        
-    }
-}
-
-
-
+//
+//protocol BinaryOperator {
+//    func andOperation()
+//    func nandOperation()
+//    func orOperation()
+//    func norOperation()
+//    func xorOperation()
+//    func notOperation()
+//    func rightShiftOperation()
+//    func leftShiftOPeration()
+//}
+//
+//class Calculator {
+//    func add(_ firstElement: String, to secondElement: String) {
+//        
+//    }
+//    
+//    func subtract(_ secondElement: String, from firstElement: String) {
+//        
+//    }
+//    
+//    func clear() {
+//        
+//    }
+//}
+//
+//class DecimalCalculator: Calculator, DecimalOperator {
+//    func multiply(_ firstElement: String, by secondElement: String) -> String {
+//        return ""
+//    }
+//    
+//    func divide() {
+//        
+//    }
+//}
+//
+//class BinaryCalculator: Calculator, BinaryOperator {
+//    func andOperation() {
+//        
+//    }
+//    
+//    func nandOperation() {
+//        
+//    }
+//    
+//    func orOperation() {
+//        
+//    }
+//    
+//    func norOperation() {
+//        
+//    }
+//    
+//    func xorOperation() {
+//        
+//    }
+//    
+//    func notOperation() {
+//        
+//    }
+//    
+//    func rightShiftOperation() {
+//        
+//    }
+//    
+//    func leftShiftOPeration() {
+//        
+//    }
+//}
+//
+//
+//
